pyproject.toml
[project]
name = "rag-basic-ai"
version = "0.1.0"
description = "Rag Basic"
authors = [
    {name = "Edwin C"}
]
license = {text = "GNU"}
readme = "docs/README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "scikit-learn (>=1.6.1,<2.0.0)",
    "pdfplumber (>=0.11.5,<0.12.0)",
    "ollama (>=0.4.7,<0.5.0)",
    "openai (>=1.66.2,<2.0.0)",
    "click (>=8.1.8,<9.0.0)",
    "rank-bm25 (>=0.2.2,<0.3.0)",
    "sentence-transformers (>=3.4.1,<4.0.0)",
    "faiss-cpu (>=1.10.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.group.dev.dependencies]
ipdb = "^0.13.13"
ipython = "^9.0.2"

